New cases added
CP condition/case does not work. Change the count values in that if statement to -1 if you wish to see the working script without the bug. Change to the listed values in the comment if you wish to see it along the rest of the program. Change all other image condition counts to -1 if you wish to see the bug case alone.